Build and Install
Create .bat builds for each set up (debug and release) and an installer.baT
Ninja is now required and it was installed. I have to install it in my other machine too.
CmakeLists files were updated, as well as build and install roots in visual studio. No need to do it in CLion. Also CLion stopped working. should try to fix that later